fix data race in when_any() with concurrent push()
In when_any(), the assumption stated in the comment that the queue.notifyerQueue_previousData
counter would no longer change, was wrong. The race takes place between a concurrent call to
when_any() and push(). If the call to setNotificationQueue() and getNotificationQueuePreviousData()
would both complete between the load of the notification queue pointer (which would still be
nullptr) and incrementing the counter, a notification would not be placed into the queue, resulting
in having permanently one data element in the queue which will only be read after the next element
has been pushed into the queue.