Fix race condition which could lead to negative overflows of the previous data counter (when_any).
It was previously thought that this race does not matter, as the counter is immediately incremented by the sender side again. The problem is,
if the sender thread sleeps long enough for the receiver thread to call when_any() in the mean time, when_any() will actually see the
negative overflow.

An assert has been added to absolutely exclude that this problem still happens and goes unnoticed.