remove GccAtomic from public dependency list
because it requires Find Module to be present and that is not installed.
But actual lib location is in linker flags anyway, so calling find_package(GccAtomic)
in implicit dependency resolution brings no benefit.
This solves a problem with non project-template based projects using
DeviceAccess, like e.g. DeviceAccess example project.